requirements.txt
python-dotenv
langchain
requests
openai
groq
streamlit